migrate ApplicationSettings to typescript
