+++ src/Settings/SettingsScreenNavigator.tsx
@@ -5,7 +5,6 @@ import {defineMessages, useIntl} from 'react-intl'
 
 import CustomPinScreen from '../../legacy/components/FirstRun/CustomPinScreen'
 import BiometricAuthScreen from '../../legacy/components/Send/BiometricAuthScreen'
-import ApplicationSettingsScreen from '../../legacy/components/Settings/ApplicationSettingsScreen'
 import BiometricsLinkScreen from '../../legacy/components/Settings/BiometricsLinkScreen'
 import ChangeCustomPinScreen from '../../legacy/components/Settings/ChangeCustomPinScreen'
 import LanguagePickerScreen from '../../legacy/components/Settings/ChangeLanguageScreen'
@@ -15,6 +14,7 @@ import ToggleEasyConfirmationScreen from '../../legacy/components/Settings/Toggl
 import {defaultNavigationOptions, defaultStackNavigatorOptions} from '../../legacy/navigationOptions'
 import {SETTINGS_ROUTES, SETTINGS_TABS} from '../../legacy/RoutesList'
 import {COLORS} from '../../legacy/styles/config'
+import {ApplicationSettingsScreen} from './ApplicationSettings'
 import {ChangePasswordScreen} from './ChangePassword/ChangePasswordScreen'
 import {ChangeWalletName} from './ChangeWalletName/ChangeWalletName'
 import {RemoveWalletScreen} from './RemoveWallet'
